:recycle:1차 클라이언트 리팩토링
