+++ client/src/pages/Join.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import "../scss/join.scss";
+import JoinFormContainer from "../components/join/JoinFormContainer";
+
+const Join: React.FC = () => {
+  return (
+    <JoinFormContainer />
+  );
+};
+
+export default Join;
