+++ client/src/pages/Login.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import "../scss/login.scss";
+import LoginFormContainer from "../components/login/LoginFormContainer";
+
+const Login: React.FC = () => {
+  return (
+    <LoginFormContainer />
+  );
+};
+
+export default Login;
