+++ client/src/pages/content/Pick.tsx
@@ -2,7 +2,6 @@ import { useState, useEffect, useRef } from "react";
 import repPick from "../../utils/pick/repPick";
 import setNumberBgColor from "../../utils/setNumberBgColor";
 import "../../scss/pick.scss";
-import Layout from "client/src/app/layout";
 
 const Pick = () => {
   const [numbers, setNumbers] = useState<number[]>([]);
@@ -35,24 +34,22 @@ const Pick = () => {
   }, [numbers]);
 
   return (
-    <Layout pageTitle="pick">
-      <div id="number">
-        <div id="number__header">
-          <header>번호 생성</header>
-        </div>
-        <div id="number__list" ref={numberListRef}>
-          {Array.from({ length: 6 }, (_, i) => (
-            <div key={i}></div>
-          ))}
-        </div>
-        <input
-          type="button"
-          value="번호 생성"
-          id="create"
-          onClick={handleClick}
-        />
+    <div id="number">
+      <div id="number__header">
+        <header>번호 생성</header>
       </div>
-    </Layout>
+      <div id="number__list" ref={numberListRef}>
+        {Array.from({ length: 6 }, (_, i) => (
+          <div key={i}></div>
+        ))}
+      </div>
+      <input
+        type="button"
+        value="번호 생성"
+        id="create"
+        onClick={handleClick}
+      />
+    </div>
   );
 };
